phase-03-programmability/requirements.txt
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
idna==3.11
invoke==2.2.1
lxml==6.0.2
ncclient==0.7.1
paramiko==4.0.0
pyang==2.7.1
pycparser==3.0
PyNaCl==1.6.2
PyYAML==6.0.3
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
xmltodict==1.0.4
